samples: fix the avr32-unknown-none sample
Bizarely enough, binutils-2.18 fails to build when CONFIG_SHELL is set
to /bin/bash (although it has code to recognise bash and act accordingly).

Waht is really strange is that it does build when CONFIG_SHELL is set
to /bin/sh, *although* /bin/sh is a symlink to /bin/bash. WTF?!?!